projects/03-ai-movie-recommender/requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# Machine learning and feature extraction
scikit-learn>=1.1.0
scipy>=1.9.0

# Text processing
nltk>=3.8
textblob>=0.17.0

# Vector database (ChromaDB)
chromadb>=0.4.0

# LangChain framework
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0

# Embeddings
openai>=1.0.0
sentence-transformers>=2.2.0

# Web framework
flask>=2.3.0
flask-cors>=4.0.0
streamlit>=1.28.0

# Environment and configuration
python-dotenv>=1.0.0

# Monitoring and logging
prometheus-client>=0.16.0

# Utilities
tqdm>=4.64.0
requests>=2.28.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0